// src/modules/error/enums/errorTypes.enum.ts
export enum ErrorTypes {
  BUILDING_NOT_FOUND = '1',
  BUILDING_LEVEL_NOT_FOUND = '2',
  BUILDING_ROOM_NOT_FOUND = '3',
  PROJECT_NOT_FOUND = '4',
  EQUIPMENT_NOT_FOUND = '5',
  ORGANIZATION_NOT_FOUND = '6',
  USER_NOT_FOUND = '7',
  DAILYCHECK_OPERATION_NOT_FOUND = '8',
  DAILYCHECK_PROCEDURE_NOT_FOUND = '9',
  UNIT_NOT_FOUND = '10',
  MAINTENANCE_OPERATION_NOT_FOUND = '11',
  MAINTENANCE_PROCEDURE_NOT_FOUND = '12',
  EQUIPMENT_CATEGORY_GROUP_NOT_FOUND = '13',
  EQUIPMENT_MODEL_NOT_FOUND = '14',
  EQUIPMENT_MANUFACTURER_NOT_FOUND = '15',
  SUBCONTRACTOR_NOT_FOUND = '16',
  COUNTRY_NOT_FOUND = '17',
  SOURCE_EQUIPMENT_NOT_FOUND = '18',
  CHAT_NOT_FOUND = '19',
  PROJECT_CATEGORY_NOT_FOUND = '20',
  STANDARD_CATEGORY_GROUP_NOT_FOUND = '21',
  BUILDING_IS_NOT_PART_OF_PROJECT = '22',
  BUILDING_LEVEL_IS_NOT_PART_OF_BUILDING = '23',
  BUILDING_ROOM_IS_NOT_PART_OF_BUILDING = '24',
  PROJECT_ID_NOT_SET = '25',
  INVALID_FILE_SIZE = '26',
  THUMBNAIL_NOT_SUPPORTED_FOR_FILE_TYPE = '27',
  PROJECT_NEEDS_START_DATE_SET = '28',
  INVALID_MAINTENANCE_PROCEDURE_ID = '29',
  INVALID_EVENT_ID = '30',
  INVALID_OPERATION_ID = '31',
  INVALID_LABEL_ID = '32',
  INVALID_PARAMETER_ID = '33',
  INVALID_DATES = '34',
  INVALID_PROJECT_ID = '35',
  USER_MUST_BE_PART_OF_ORGANIZATION = '36',
  INVALID_ORGANIZATION_TYPE = '37',
  USER_DOES_NOT_HAVE_A_VALID_ROLE = '38',
  USER_TO_ASSIGN_NOT_FOUND = '39',
  COULD_NOT_SEND_EMAIL = '40',
  CHAT_DOES_NOT_EXIST_OR_USER_IS_NOT_A_PARTICIPANT = '41',
  YOU_CAN_ADD_LABELS_ONLY_TO_VISUAL_OPERATION = '42',
  YOU_CAN_ADD_PARAMETERS_ONLY_TO_PARAMETER_OPERATION = '43',
  YOU_NEED_TO_SET_EITHER_EQUIPMENT_OR_EQUIPMENT_CATEGORY_GROUP_FOR_DAILYCHECK_PROCEDURE = '44',
  PARENT_ORGANIZATION_TYPE_MUST_BE_PROVIDER = '45',
  YOU_NEDD_TO_SET_EITHER_A_LABEL_OR_A_PARAMETER = '46',
  USER_TO_ASIGN_IS_NOT_PART_OF_AN_ORGANIZATION = '47',
  YOU_CAN_ACTIVATE_ONLY_PROJECTS_IN_DRAFT_STATE = '48',
  INITIATOR_USER_NOT_FOUND = '49',
  UNABLE_TO_DELETE_BUILDING = '50',
  UNABLE_TO_DELETE_BUILDING_FK_CONSTRAINT = '51',
  UNABLE_TO_DELETE_BUILDING_LEVEL = '52',
  UNABLE_TO_DELETE_BUILDING_LEVEL_FK_CONSTRAINT = '53',
  UNABLE_TO_DELETE_BUILDING_ROOM = '54',
  UNABLE_TO_DELETE_BUILDING_ROOM_FK_CONSTRAINT = '55',
  UNABLE_TO_DELETE_ORGANIZATION = '56',
  UNABLE_TO_DELETE_ORGANIZATION_FK_CONSTRAINT = '57',
  UNABLE_TO_DELETE_SUBCONTRACTOR = '58',
  UNABLE_TO_DELETE_SUBCONTRACTOR_FK_CONSTRAINT = '59',
  UNABLE_TO_DELETE_USER = '60',
  UNABLE_TO_DELETE_USER_FK_CONSTRAINT = '61',
  UNABLE_TO_DELETE_EQUIPMENT = '62',
  UNABLE_TO_DELETE_EQUIPMENT_FK_CONSTRAINT = '63',
  UNABLE_TO_DELETE_EQUIPMENT_NOT_DELETABLE = '64',
  UNABLE_TO_DELETE_EQUIPMENT_READ_ONLY = '65',
  EQUIPMENT_IS_READ_ONLY = '66',
  UNABLE_TO_DELETE_PROJECT = '67',
  UNABLE_TO_DELETE_PROJECT_FK_CONSTRAINT = '68',
  USER_NOT_FOUND_IN_PROJECT = '69',
  TICKET_UPDATE_NEED_TO_SPECIFY_DUE_DATE_IF_CHANGING_PRIORITY = '70',
  TICKET_NOT_FOUND = '71',
  TICKET_RECIPIENT_MUST_BE_SET = '72',
  INTERNAL_CHAT_IS_ONY_FOR_PROVIDER_USERS = '73',
  USER_ALREADY_ACTIVE_IN_CHAT = '74',
  USER_NOT_ACTIVE_IN_CHAT = '75',
  EQUIPMENT_CATEGORY_NOT_FOUND = '76',
  PROCUREMENT_NOT_FOUND = '77',
  FILE_NOT_FOUND = '78',
  EMAIL_IN_USE = '79',
  DAILYCHECK_GROUP_NOT_FOUND = '80',
  DAILYCHECK_GROUP_EXISTS = '81',
  WORK_ORDER_NOT_FOUND = '82',
  TICKET_INVALID_ORGANIZATION_TYPE = '83',
  RESTRICTED_FROM_EXTERNAL_CHAT = '84',
  YOU_NEED_TO_HAVE_AT_LEAST_ONE_USER_FROM_THE_RECIPIENT_ORG_IN_CHAT = '85',
  TICKET_STATUS_CHANGE_INVALID = '86',
  USER_NOT_IN_ORGANIZATION = '87',
  DC_EVENT_NOT_FOUND = '88',
  PERSONAL_AGENDA_DOES_NOT_EXIST = '89',
  PERSONAL_AGENDA_REMINDER_DATE_MUST_BE_BEFORE_START_DATE = '90',
  PERSONAL_AGENDA_MUST_RECEIVE_BOTH_START_AND_END_DATES = '91',
  PERSONAL_AGENDA_START_DATE_MUST_BE_BEFORE_END_DATE = '92',
  UNABLE_TO_DELETE_PERSONAL_AGENDA = '93',
  DAILYCHECK_NOT_FOUND = '94',
  STANDARD_CATEGORY_GROUP_EXISTS = '95',
  STANDARD_CATEGORY_NOT_FOUND = '96',
  STANDARD_CATEGORY_EXISTS = '97',
  STANDARD_PROCEDURE_NOT_FOUND = '98',
  STANDARD_OPERATION_NOT_FOUND = '99',
  STANDARD_PROJECT_CATEGORY_EXISTS = '100',
  PMP_EVENT_NOT_FOUND = '101',
  PROCUREMENT_INVALID_ORGANIZATION_TYPE = '102',
  CHAT_MESSAGE_NOT_FOUND = '103',
  CHAT_MESSAGE_AUTH_USER_NOT_OWNER_OF_MESSAGE = '104',
}